requirements.txt
aioamqp==0.14.0
aiohttp==3.7.4
aiormq==3.3.0
async-timeout==3.0.1
attrs==20.2.0
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
idna==2.10
IMAPClient==2.1.0
multidict==4.7.6
pamqp==2.3.0
pika==1.1.0
pycodestyle==2.6.0
python-dotenv==0.18.0
pyzmail36==1.0.4
requests==2.24.0
selenium==3.141.0
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.25.10
yarl==1.5.1
